Move 'viewer' stuff into stats
these screens will have to be merged together into some kind of nice pager/UI when I start adding graphs
