+++ app/src/main/java/com/github/saikcaskey/pokertracker/ui/root/RootComponent.kt
@@ -1,7 +1,19 @@
-package com.github.saikcaskey.pokertracker.domain.components
+package com.github.saikcaskey.pokertracker.ui.root
 
 import com.arkivanov.decompose.router.stack.ChildStack
 import com.arkivanov.decompose.value.Value
+import com.github.saikcaskey.pokertracker.domain.components.AccountFeatureComponent
+import com.github.saikcaskey.pokertracker.domain.components.EventDetailComponent
+import com.github.saikcaskey.pokertracker.domain.components.ExpenseDetailComponent
+import com.github.saikcaskey.pokertracker.domain.components.InsertEventComponent
+import com.github.saikcaskey.pokertracker.domain.components.InsertExpenseComponent
+import com.github.saikcaskey.pokertracker.domain.components.InsertVenueComponent
+import com.github.saikcaskey.pokertracker.domain.components.MainComponent
+import com.github.saikcaskey.pokertracker.domain.components.PlannerDayDetailComponent
+import com.github.saikcaskey.pokertracker.domain.components.VenueDetailComponent
+import com.github.saikcaskey.stats.domain.ViewEventsComponent
+import com.github.saikcaskey.stats.domain.ViewExpensesComponent
+import com.github.saikcaskey.stats.domain.ViewVenuesComponent
 
 /**
  * The navigation "Root" for the app's decompose structure.
