+++ app/src/main/java/com/github/saikcaskey/pokertracker/ui/root/RootContent.kt
@@ -8,17 +8,17 @@ import com.arkivanov.decompose.extensions.compose.stack.Children
 import com.arkivanov.decompose.extensions.compose.stack.animation.*
 import com.github.saikcaskey.pokertracker.ui_compose.components.venue.InsertVenueContent
 import com.github.saikcaskey.pokertracker.ui_compose.components.venue.VenueDetailContent
-import com.github.saikcaskey.pokertracker.ui_compose.components.venue.ViewVenuesContent
-import com.github.saikcaskey.pokertracker.domain.components.RootComponent
-import com.github.saikcaskey.pokertracker.domain.components.RootComponent.Child
+import com.github.saikcaskey.stats.presentation.ViewVenuesContent
+import com.github.saikcaskey.pokertracker.ui.root.RootComponent
+import com.github.saikcaskey.pokertracker.ui.root.RootComponent.Child
 import com.github.saikcaskey.pokertracker.ui.main.MainContent
 import com.github.saikcaskey.pokertracker.ui_compose.common.theme.AppTheme
 import com.github.saikcaskey.pokertracker.ui_compose.components.event.EventDetailContent
 import com.github.saikcaskey.pokertracker.ui_compose.components.event.InsertEventContent
-import com.github.saikcaskey.pokertracker.ui_compose.components.event.ViewEventsContent
+import com.github.saikcaskey.stats.presentation.ViewEventsContent
 import com.github.saikcaskey.pokertracker.ui_compose.components.expense.ExpenseDetailContent
 import com.github.saikcaskey.pokertracker.ui_compose.components.expense.InsertExpenseContent
-import com.github.saikcaskey.pokertracker.ui_compose.components.expense.ViewExpensesContent
+import com.github.saikcaskey.stats.presentation.ViewExpensesContent
 import com.github.saikcaskey.pokertracker.planner.composables.PlannerDayDetailContent
 import com.github.saikcaskey.account.presentation.AccountFeatureContent
 
@@ -27,7 +27,7 @@ fun RootContent(
     component: RootComponent,
     modifier: Modifier = Modifier,
 ) {
-    AppTheme() {
+    AppTheme {
         Surface(
             modifier = modifier.fillMaxSize()
         ) {
