+++ feature/stats/src/main/java/com/github/saikcaskey/stats/data/DefaultViewExpensesComponent.kt
@@ -1,13 +1,13 @@
-package com.github.saikcaskey.pokertracker.presentation.expense
+package com.github.saikcaskey.stats.data
 
 import com.arkivanov.decompose.ComponentContext
 import com.github.saikcaskey.pokertracker.domain.CoroutineDispatchers
-import com.github.saikcaskey.pokertracker.domain.components.ViewExpensesComponent
+import com.github.saikcaskey.stats.domain.ViewExpensesComponent
 import com.github.saikcaskey.pokertracker.domain.models.Expense
 import com.github.saikcaskey.pokertracker.domain.models.adjustedAmount
 import com.github.saikcaskey.pokertracker.domain.repository.ExpenseRepository
-import com.github.saikcaskey.pokertracker.domain.components.ViewExpensesComponent.*
-import com.github.saikcaskey.pokertracker.domain.components.ViewExpensesComponent.ExpenseSortOption.*
+import com.github.saikcaskey.stats.domain.ViewExpensesComponent.*
+import com.github.saikcaskey.stats.domain.ViewExpensesComponent.ExpenseSortOption.*
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.flow.SharingStarted.Companion.Eagerly
