+++ feature/stats/src/main/java/com/github/saikcaskey/stats/data/DefaultViewVenuesComponent.kt
@@ -1,11 +1,11 @@
-package com.github.saikcaskey.pokertracker.presentation.venue
+package com.github.saikcaskey.stats.data
 
 import com.arkivanov.decompose.ComponentContext
 import com.github.saikcaskey.pokertracker.domain.CoroutineDispatchers
 import com.github.saikcaskey.pokertracker.domain.models.Venue
-import com.github.saikcaskey.pokertracker.domain.components.ViewVenuesComponent
+import com.github.saikcaskey.stats.domain.ViewVenuesComponent
 import com.github.saikcaskey.pokertracker.domain.repository.VenueRepository
-import com.github.saikcaskey.pokertracker.domain.components.ViewVenuesComponent.*
+import com.github.saikcaskey.stats.domain.ViewVenuesComponent.*
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.flow.SharingStarted.Companion.Eagerly
