+++ feature/stats/src/main/java/com/github/saikcaskey/stats/presentation/ViewEventsContent.kt
@@ -1,4 +1,4 @@
-package com.github.saikcaskey.pokertracker.ui_compose.components.expense
+package com.github.saikcaskey.stats.presentation
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
@@ -7,22 +7,21 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.github.saikcaskey.pokertracker.domain.components.ViewExpensesComponent
 import com.github.saikcaskey.pokertracker.ui_compose.common.appbar.TopAppBarItemViewer
 import com.github.saikcaskey.pokertracker.ui_compose.common.inputform.InputDropdownSimple
-import com.github.saikcaskey.pokertracker.ui_compose.common.profitsummary.AnimatedExpenseText
+import com.github.saikcaskey.stats.domain.ViewEventsComponent
 
 @Composable
-fun ViewExpensesContent(component: ViewExpensesComponent) {
+fun ViewEventsContent(component: ViewEventsComponent) {
     val uiState by component.uiState.collectAsState()
 
     Scaffold(
         topBar = {
             TopAppBarItemViewer(
-                "All Expenses",
+                "All Events",
                 onBackClicked = component::onBackClicked,
-                onShowInsertItemClicked = component::onShowInsertExpenseClicked,
-                onDeleteAllItemsClicked = component::onDeleteAllExpensesClicked,
+                onShowInsertItemClicked = component::onShowInsertEventClicked,
+                onDeleteAllItemsClicked = component::onDeleteAllEventsClicked,
             )
         },
     ) { scaffoldPadding ->
@@ -39,19 +38,20 @@ fun ViewExpensesContent(component: ViewExpensesComponent) {
                 modifier = Modifier.fillMaxWidth()
             )
             Spacer(modifier = Modifier.height(8.dp))
+
             InputDropdownSimple(
                 selected = uiState.searchFilter.sort,
-                entries = ViewExpensesComponent.ExpenseSortOption.entries,
+                entries = ViewEventsComponent.EventSortOption.entries,
                 onSelected = component::onFilterOptionChanged
             )
             Spacer(modifier = Modifier.height(8.dp))
 
             LazyColumn {
-                items(uiState.filtered) { expense ->
+                items(uiState.filtered) { event ->
                     ListItem(
-                        headlineContent = { AnimatedExpenseText(expense) },
-                        supportingContent = { Text(expense.createdAt.toString()) },
-                        modifier = Modifier.clickable { component.onShowExpenseDetailClicked(expense.id) }
+                        headlineContent = { Text(event.name ?: event.id.toString()) },
+                        supportingContent = { Text(event.createdAt?.toString().orEmpty()) },
+                        modifier = Modifier.clickable { component.onShowEventDetailClicked(event.id) }
                     )
                     HorizontalDivider()
                 }
